+++ inferno-ml-server/src/Inferno/ML/Server/Inference/Model.hs
@@ -3,33 +3,27 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Inferno.ML.Server.Inference.Model
-  ( getModelsAndVersions,
-    getTorchScriptModelContents,
-    cacheAndLoadModel,
-    mkModelPath,
+  ( loadModel,
   )
 where
 
 import Control.Arrow ((&&&))
-import Control.Monad ((<=<))
 import Control.Monad.Extra (loopM, unlessM, whenM)
 import Control.Monad.IO.Class (MonadIO (liftIO))
 import Control.Monad.ListM (sortByM)
 import Data.ByteString (ByteString)
-import Data.Foldable (toList)
 import qualified Data.Text as Text
 import Data.UUID (UUID)
 import qualified Data.UUID as UUID
-import Data.Vector (Vector)
 import Data.Word (Word64)
 import Database.PostgreSQL.Simple
-  ( In (In),
-    Only (Only),
+  ( Only (Only),
     Query,
     withTransaction,
   )
 import Database.PostgreSQL.Simple.LargeObjects
   ( IOMode (ReadMode),
+    Oid,
     loClose,
     loOpen,
     loRead,
@@ -52,92 +46,49 @@ import UnliftIO.Directory
 import UnliftIO.Exception (bracket, catchIO, displayException)
 import UnliftIO.IO.File (writeBinaryFileDurableAtomic)
 
--- | Get an array of model versions along with the parent model of each; note
--- that this does not retrieve the model version contents -- each version only
--- contains the 'Oid' of the large object
-getModelsAndVersions ::
-  Vector (Id ModelVersion) -> RemoteM (Vector (Model, ModelVersion))
-getModelsAndVersions =
-  fmap (fmap joinToTuple)
-    . queryStore q
-    . Only
-    . In
-    . toList
-  where
-    q :: Query
-    q =
-      [sql|
-        SELECT M.*, V.*
-        FROM mversions V
-          INNER JOIN models M ON V.model = M.id
-        WHERE V.id IN ?
-          AND V.terminated IS NULL
-          AND M.terminated IS NULL
-      |]
-
--- | Get the actual serialized bytes of a TorchScript model, which is stored in
--- the Postgres large object table (and must be explicitly imported using 'loImport'),
--- along with the number of bytes
---
--- TODO/NOTE This will ONLY be invoked in the future when a model has a @TorchScript@
--- @ModelConfig@. At the moment, it is invoked for all models. This is the least
--- intrusive change (i.e. throwing an exception if a @Bedrock@ model sneaks in).
--- We have no path to creating @Bedrock@ at the moment (except via direct database
--- manipulation), so this is reasonable for now
-getTorchScriptModelContents :: ModelVersion -> RemoteM ByteString
-getTorchScriptModelContents mversion =
-  case mversion.contents of
-    Bedrock _ ->
-      throwRemoteError . OtherRemoteError $
-        Text.unwords
-          [ "Model"
-          , tshow mversion.id
-          , "is a Bedrock model; cannot be used as a TorchScript model"
-          ]
-    TorchScript oid ->
-      withConns $ \conn -> withRunInIO $ \r ->
-        withTransaction conn . r $ do
-          liftIO
-            . bracket (loOpen conn oid ReadMode) (loClose conn)
-            . flip (loRead conn)
-            $ fromIntegral mversion.size
+-- | Load either a @TorchScript@ model, which will also be cached in the
+-- model directory of the server, or a @Bedrock@ model, in which case the
+-- model configuration is simply held in memory
+loadModel :: UUID -> RemoteM (ModelConfig ScriptModule)
+loadModel uuid =
+  getModelVersion uuid >>= \mversion ->
+    case mversion.contents of
+      TorchScript oid ->
+        fmap TorchScript . cacheTorchScriptModel $
+          TorchScriptModel uuid mversion.size oid
+      -- For `Bedrock`, we can just use the configuration straight from the DB
+      Bedrock cfg -> pure $ Bedrock cfg
 
 -- | Cache and load a model by its UUID. If already cached, loads directly.
 -- Otherwise fetches from DB, checks disk space (evicting LRU models if needed),
--- writes to cache, then loads
-cacheAndLoadModel :: UUID -> RemoteM ScriptModule
-cacheAndLoadModel uuid = do
+-- writes to cache, then loads into a Torch @ScriptModule@
+--
+-- NOTE: This only applies to @TorchScript@ models, which have actual contents
+-- (the blob of the @.ts.pt@); @Bedrock@ models are loaded differently
+cacheTorchScriptModel :: TorchScriptModel -> RemoteM ScriptModule
+cacheTorchScriptModel tsm = do
   unlessM (doesPathExist path) $ do
-    logInfo . CopyingModel $ Id uuid
-    copyModelVersion =<< getModelVersion
+    logInfo . CopyingModel $ Id tsm.id
+    checkDiskSpace tsm.size
+    writeBinaryFileDurableAtomic path =<< getTorchScriptModelContents
   -- NOTE: Any exceptions from `loadScript` are caught in the `loadModel`
   -- primitive and converted to `RuntimeError`s there
   liftIO $ Torch.Script.loadScript WithoutRequiredGrad path
   where
     path :: FilePath
-    path = modelCachePath </> mkModelPath uuid
+    path = modelCachePath </> UUID.toString tsm.id <.> "ts" <.> "pt"
 
-    copyModelVersion :: ModelVersion -> RemoteM ()
-    copyModelVersion mversion = do
-      checkDiskSpace mversion.size
-      writeBinaryFileDurableAtomic path
-        =<< getTorchScriptModelContents mversion
-
-    getModelVersion :: RemoteM ModelVersion
-    getModelVersion =
-      firstOrThrow (NoSuchModel (Right (Id uuid)))
-        =<< queryStore q (Only uuid)
-      where
-        q :: Query
-        q =
-          [sql|
-            SELECT V.*
-            FROM mversions V
-              INNER JOIN models M ON V.model = M.id
-            WHERE V.id = ?
-              AND V.terminated IS NULL
-              AND M.terminated IS NULL
-          |]
+    -- Get the actual serialized bytes of a TorchScript model, which is stored in
+    -- the Postgres large object table (and must be explicitly imported using 'loImport'),
+    -- along with the number of bytes
+    getTorchScriptModelContents :: RemoteM ByteString
+    getTorchScriptModelContents =
+      withConns $ \conn -> withRunInIO $ \r ->
+        withTransaction conn . r
+          $ liftIO
+            . bracket (loOpen conn tsm.oid ReadMode) (loClose conn)
+            . flip (loRead conn)
+          $ fromIntegral tsm.size
 
     -- Check if adding a model would exceed 90% disk usage; if so, evict LRU models
     checkDiskSpace :: Word64 -> RemoteM ()
@@ -145,10 +96,22 @@ cacheAndLoadModel uuid = do
       flip whenM evictOldModels . wouldExceedThreshold $ fromIntegral modelSize
 
     evictOldModels :: RemoteM ()
-    evictOldModels = loopM doEvict =<< modelsByAccessTime modelCachePath
+    evictOldModels = loopM doEvict =<< modelsByAccessTime
+      where
+        -- Get cached `TorchScript` models sorted by access time (LRU first)
+        modelsByAccessTime :: RemoteM [FilePath]
+        modelsByAccessTime =
+          sortByM cmpAccessTime
+            =<< listDirectory modelCachePath
+          where
+            cmpAccessTime :: FilePath -> FilePath -> RemoteM Ordering
+            cmpAccessTime f1 f2 =
+              compare <$> getAccessTime f1 <*> getAccessTime f2
 
     doEvict :: [FilePath] -> RemoteM (Either [FilePath] ())
     doEvict = \case
+      -- Either the dir started empty or we've deleted all remaining old cached
+      -- model versions in this case
       [] -> pure $ Right ()
       m : ms ->
         wouldExceedThreshold 0 >>= \case
@@ -172,13 +135,28 @@ cacheAndLoadModel uuid = do
         threshold :: DiskUsage -> Integer
         threshold usage = (fromIntegral usage.diskTotal * 9) `div` 10
 
--- | Get models sorted by access time (least recently used first)
-modelsByAccessTime :: (MonadIO m) => FilePath -> m [FilePath]
-modelsByAccessTime = sortByM cmpAccessTime <=< listDirectory
+getModelVersion :: UUID -> RemoteM ModelVersion
+getModelVersion uuid =
+  firstOrThrow (NoSuchModel (Right (Id uuid)))
+    =<< queryStore q (Only uuid)
   where
-    cmpAccessTime :: (MonadIO m) => FilePath -> FilePath -> m Ordering
-    cmpAccessTime f1 f2 = compare <$> getAccessTime f1 <*> getAccessTime f2
+    q :: Query
+    q =
+      [sql|
+        SELECT V.*
+        FROM mversions V
+          INNER JOIN models M ON V.model = M.id
+        WHERE V.id = ?
+          AND V.terminated IS NULL
+          AND M.terminated IS NULL
+      |]
 
--- | Generate a filepath from a model version UUID
-mkModelPath :: UUID -> FilePath
-mkModelPath = (<.> "ts" <.> "pt") . UUID.toString
+-- Local helper for the contents of a `TorchScript` model; this is after we have
+-- retrieved the `ModelVersion` from the database; we don't want to keep
+-- pattern-matching on the `contents` type, so so this bundles the needed
+-- information for loading and caching the `TorchScript` model
+data TorchScriptModel = TorchScriptModel
+  { id :: UUID
+  , size :: Word64
+  , oid :: Oid
+  }